% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_a.R
\docType{data}
\name{data_a}
\alias{data_a}
\title{data_a - Age and Looks}
\format{
A data frame with 40 rows and 6 variables:
\describe{
\item{bsex}{biological sex}
\item{age}{age of person}
\item{height}{height of person}
\item{hair}{hair colour}
\item{eyes}{eye colour}
\item{glasses}{info on whether person wears glasses}
}
}
\source{
Schmitz, Dominic & Esser, Janina. (2021). SfL: Statistics for Linguistics. URL: https://github.com/dosc91/SfL
}
\usage{
data_a
}
\description{
A dataset containing fictional data on age and looks.
Q1: Is there are dependency between different variables?
}
\keyword{datasets}
