% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_c.R
\docType{data}
\name{data_c}
\alias{data_c}
\title{data_c - Incomplete Neutralization}
\format{
A data frame with 300 rows and 9 variables:
\describe{
\item{subject}{biological sex}
\item{gender}{gender of subject}
\item{item}{word produced}
\item{voicing}{voicing of final consonant}
\item{vowel}{vowel preceding the final consonant}
\item{place}{place of articulation of final consonants}
\item{accent_type}{nuclear or prenuclear accent on word}
\item{prosodic_boundary}{info on whether word is followed by a boundary}
\item{vowel_dur}{duration of vowel preceding the final consonant}
}
Roettger, T. B., Winter, B., Kirby, J., Grawunder, S., & Grice, M. (2014).
Assessing incomplete neutralization of final devoicing in German. Journal of Phonetics, 43: 11- 25.
}
\usage{
data_c
}
\description{
The Incomplete Neutralization data set, data_c, contains data on the phonological phenomenon of incomplete neutralization. The values are fictional,
but closely related to real data, i.e. the data by Roettger et al. (2014).
A possible question to investigate with this data set is whether vowels are shorter when preceding voiceless word-final consonants.
}
\keyword{datasets}
