% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_p.R
\docType{data}
\name{data_p}
\alias{data_p}
\title{data_p - Program Awards}
\format{
A data frame with 200 rows and 4 variables:
\describe{
\item{id}{ID}
\item{num_awards}{number of awards}
\item{prog}{program}
\item{math}{math grade}
}
}
\source{
URL: https://stats.idre.ucla.edu/stat/data/poisson_sim.csv
}
\usage{
data(data_p)
}
\description{
The Program Awards data set, data_p, contains data on fictional awards won by pupils in different math programs.
}
\keyword{datasets}
