% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_t.R
\docType{data}
\name{data_t}
\alias{data_t}
\title{data_t - Tonal Alignment in German}
\format{
A data frame with 544 rows and 13 variables:
\describe{
\item{speaker}{speaker ID}
\item{word_dur}{word duration}
\item{vowel}{vowel of word}
\item{sylnum}{number of syllables of word}
\item{sylstr}{syllable structure of first syllable of word}
\item{C1_dur}{duration of consonant preceding the vowel}
\item{V1_dur}{vowel duration of first syllable vowel}
\item{C2_dur}{duration of first consonant following the vowel within the same syllable}
\item{C3_dur}{duration of second consonant following the vowel within the same syllable}
\item{W_dur}{duration of onset consonant of 2nd syllable}
\item{A1_dur}{duration of vowel of 2nd syllable}
\item{L_dur}{duration of onset consonant of 3rd syllable}
\item{A2_dur}{duration of vowel of 3rd syllable}
\item{L_time}{point in time when lowest F0 value, i.e. L, occured}
\item{L_freq}{frequency of lowest F0 value, i.e. L}
}
}
\source{
This data set is a shortened version of: Schmitz, D. (2018). Tonal Alignment and Segmental Composition in German.
Master's Thesis. Universität zu Köln, Köln, Germany.
}
\usage{
data_t
}
\description{
The Tonal Alignment in German data set, data_t, contains data on tonal alignment in German. The values are fictional, but closely
related to real data, i.e. the data by Schmitz (2018).
A possible question to investigate with this data set is whether the frequency of low points differs by syllable structure.
}
\keyword{datasets}
