% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_z.R
\docType{data}
\name{data_z}
\alias{data_z}
\title{data_z - Highly Correlated Data}
\format{
A data frame with 653 rows and 14 variables:
\describe{
\item{Word}{word}
\item{Affix}{type of /s/}
\item{sDur}{/s/ duration}
\item{sDurLog}{logged /s/ duration}
\item{speechrate_1}{speech rate}
\item{speechrate_2}{speech rate}
\item{speechrate_3}{speech rate}
\item{wordDur}{word duration}
\item{baseDur}{base duration}
\item{frequency1}{frequency}
\item{frequency2}{frequency}
\item{frequency3}{frequency}
\item{googleFrequency}{frequency via Google}
\item{reactionTime}{reaction time (RT)}
}
}
\source{
Schmitz, D., & Esser, J. (2021). SfL: Statistics for Linguistics. R package version 0.4. URL: https://github.com/dosc91/SfL
}
\usage{
data(data_z)
}
\description{
The Highly Correlated Data data set, data_z, contains highly correlated data on /s/ durations.
}
\keyword{datasets}
