% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_markdown.R
\name{load_markdown}
\alias{load_markdown}
\title{Open markdown file.}
\usage{
load_markdown(session)
}
\arguments{
\item{session}{Session number as a number.}
}
\value{
Nothing. Used for its side effects.
}
\description{
Call this function with the session number as argument and the corresponding RMarkdown
file will be copied in the working directory and opened in the editor.
}
\details{
The following sessions have markdown files:
\itemize{
\item Session 01.
}
}
