% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_exercises.R
\name{open_exercises}
\alias{open_exercises}
\title{Open Exercise}
\usage{
\code{open_exercise(session_number)}
}
\arguments{
\item{session_number}{The number of a session, i.e. 02, 03, 04, 05, 06, or 07.}
}
\value{
Nothing.
}
\description{
Calling \code{open_exercise} with a session number as argument will open the exercise html for that session.
The html will be opened in your default browser. This function is inspired by Coretta et al. (see references).
}
\examples{
open_exercises(02)

}
\references{
Coretta, Stefano, Casillas, Joseph V., and Roettger, Timo. (2021). learnB4SS: Learning materials for the learnB4SS workshop. R package version 1.0.0. https://github.com/learnB4SS/learnB4SS
}
\author{
D. Schmitz
}
