% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_slides.R
\name{open_slides}
\alias{open_slides}
\title{Open Slides}
\usage{
\code{open_slides(session_number)}
}
\arguments{
\item{session_number}{The number of a session, i.e. 01, 02, 03, 04, 05, 06, or 07.}
}
\value{
Nothing.
}
\description{
Calling \code{open_slides} with a session number as argument will open the powerpoint slides for that session as pdf.
The pdf will be opened in your default browser. This function is inspired by Coretta et al. (see references).
}
\examples{
open_slides(02)

}
\references{
Coretta, Stefano, Casillas, Joseph V., and Roettger, Timo. (2021). learnB4SS: Learning materials for the learnB4SS workshop. R package version 1.0.0. https://github.com/learnB4SS/learnB4SS
}
\author{
D. Schmitz
}
