% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordinal_p.R
\name{ordinal_p}
\alias{ordinal_p}
\title{Compute p-values for ordinal regression}
\usage{
ordinal_p(mdl)
}
\arguments{
\item{mdl}{A fitted \code{polr} model.}
}
\value{
A coefficient matrix with an additional \verb{p value} column.
}
\description{
Adds two-sided Wald p-values to the coefficient table of a model fitted with
\code{\link[MASS:polr]{MASS::polr()}}.
}
\examples{
\dontrun{
model <- MASS::polr(Sat ~ Infl + Type + Cont, MASS::housing, Hess = TRUE)
ordinal_p(model)
}
}
