% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se.R
\name{se}
\alias{se}
\title{Standard error}
\usage{
se(x, na.rm = FALSE)
}
\arguments{
\item{x}{A numeric vector.}

\item{na.rm}{Remove missing values before computing the standard error.}
}
\value{
A numeric scalar.
}
\description{
Computes the standard error of the mean.
}
\examples{
se(c(1, 2, 3, 4))
}
