% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corfun.R
\name{corfun}
\alias{corfun}
\title{Compute Correlation Coefficients}
\usage{
corfun(data, variables)
}
\arguments{
\item{data}{The data set containing the variables.}

\item{variables}{The variables for which correlation coefficients are to be computed.}
}
\description{
\code{corfun} computes the Pearson or Spearman correlation coefficients for all variable pairs. Pearson is computed for comparisons of
normally distributed numeric variables, Spearman for all other comparisons.
}
\examples{
data("data_s")

corfun(data_s, c("sDurLog", "speakingRate", "typeOfS"))

corfun(data_s, c("sDurLog", "speakingRate", "typeOfS", "sDur"))

corfun(data_s, c("sDurLog", "speakingRate", "typeOfS", "sDur", "pauseDur", "pauseBin"))

}
\author{
D. Schmitz
}
