% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_matrix.R
\name{correlation_matrix}
\alias{correlation_matrix}
\title{Create Correlation Matrix}
\usage{
\code{correlation_matrix(data = , variables = c(""))}
}
\arguments{
\item{data}{The data set containing the variables to be plotted.}

\item{variables}{The variables used in the plot. Note: Variables cannot be of type 'character'.}
}
\value{
Nothing. Creates a plot.
}
\description{
This is a wrapper for \code{languageR::pairscor.fnc}. A matrix of scatter plots is produced with
Pearson and Spearman correlations in the lower triangle. Fall back to the original function to specify additional arguments.
}
\examples{
data("data_s")

correlation_matrix(data_s, variables = c("typeOfS", "sDur"))

}
\references{
R. Harald Baayen and Elnaz Shafaei-Bajestan (2019). languageR: Analyzing Linguistic Data: A Practical Introduction to Statistics.
R package version 1.5.0. https://CRAN.R-project.org/package=languageR
}
\author{
D. Schmitz
}
