% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_error_bar_df.R
\name{create_error_bar_df}
\alias{create_error_bar_df}
\title{Create Error Bar Dataframe}
\usage{
\code{create_error_bar_df(data, numerical, factors)}
}
\arguments{
\item{data}{The original data set.}

\item{numerical}{The numerical variable which will be visualised in the bar plot.}

\item{factors}{}
}
\value{
A dataframe object.
}
\description{
Create a dataframe with which a \code{ggplot2} bar plot with error bars (+/- 1 standard deviation) can be created.
}
\author{
D. Schmitz
}
