% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_a.R
\docType{data}
\name{data_a}
\alias{data_a}
\title{data_a - Age and Looks}
\format{
A data frame with 40 rows and 6 variables:
\describe{
\item{bsex}{biological sex}
\item{age}{age of person}
\item{height}{height of person}
\item{hair}{hair colour}
\item{eyes}{eye colour}
\item{glasses}{info on whether person wears glasses}
}
}
\source{
Schmitz, D., & Esser, J. (2022). SfL: Statistics for Linguistics. R package version 0.3. URL: https://github.com/dosc91/SfL
}
\usage{
data(data_a)
}
\description{
The Age and Looks data set, data_a, contains data on age and appearance of fictional individuals.
A possible question to investigate with this data set is whether there is a dependency between different variables of the data set.
}
\keyword{datasets}
