% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_s.R
\docType{data}
\name{data_s}
\alias{data_s}
\title{data_s - Duration of word-final /s/ in English}
\format{
A data frame with 150 rows and 24 variables:
\describe{
\item{number}{unique ID}
\item{item}{pseudoword produced, contains the /s/}
\item{speaker}{speaker ID}
\item{bsex}{biological sex of speaker}
\item{list}{ID of pseudo-randomized list}
\item{typeOfS}{type of /s/, i.e. non-morphemic, plural, is-clitic}
\item{folSeg}{phone following the /s/}
\item{slideNumber}{number of slide during experiment}
\item{transcription}{transcription of item}
\item{sDur}{duration of word-final /s/ in seconds}
\item{baseDur}{duration of string of segments before /s/ in seconds}
\item{sDurLog}{log-transformed sDur}
\item{baseDurLog}{log-transformed baseDur}
\item{biphoneProbSum}{summed biphone probability of transcription}
\item{biphoneProbSumBin}{binary version of biphoneProbSum}
\item{age}{age of speaker}
\item{location}{home of speaker}
\item{monoMultilingual}{mono- vs multilinguality of speaker}
\item{speakingRate}{speech rate}
\item{pauseDur}{duration of pause following the /s/}
\item{pauseBin}{binary version of pauseDur}
\item{folType}{segmental type of folSeg}
\item{preC}{phone preceding the /s/}
\item{googleFreq}{number of Google hits for item}
}
}
\source{
Schmitz, D., Baer-Henney, D., & Plag, I. (2021). The duration of word-final /s/ differs across morphological categories in English: evidence from pseudowords. Phonetica. De Gruyter Mouton 78(5–6). 571–616. https://doi.org/10.1515/phon-2021-2013.
}
\usage{
data_s
}
\description{
The Duration of word-final /s/ in English data set, data_s, contains data on three different types of word-final /s/ in
English, i.e. non-morphemic, plural, and is-clitic /s/. The data is fictional, but closely related to the original data used in a study by Schmitz et al. (2021).
A possible question to investigate with this data set is whether the duration of /s/ depends on the type of /s/.
}
\keyword{datasets}
