% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_v.R
\docType{data}
\name{data_v}
\alias{data_v}
\title{data_v - Vowel Shortening in German}
\format{
A data frame with 448 rows and 6 variables:
\describe{
\item{speaker}{speaker ID}
\item{word}{pseudoword produced, contains the vowel}
\item{duration}{vowel duration}
\item{structure}{coda structure}
\item{vowel}{one of five vowels}
\item{rate}{speech rate}
}
}
\source{
Schmitz, D., Hae-Eun C., & Niemann, H. (2018). Vowel shortening in German as a function of syllable structure. In Proceedings 13. Phonetik und Phonologie Tagung (P&P13), 181–184. Berlin.
}
\usage{
data_v
}
\description{
The Vowel Shortening in German data set, data_v, contains data on the duration of vowels in open, simplex, and complex syllables
in German. The data is fictional, but closely related to the original data used by Schmitz et al. (2018).
A possible question to investigate with this data set is whether vowel duration differs by syllable structure.
}
\keyword{datasets}
