% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heteroscedasticity_check.R
\name{heteroscedasticity_check}
\alias{heteroscedasticity_check}
\title{Check Heteroscedasticity and Linearity of a Model}
\usage{
heteroscedasticity_check(mdl, random = NULL)
}
\arguments{
\item{mdl}{A simple or multiple linear regression model.}
}
\value{
A plot and a test results in the console.
}
\description{
Checks a simple or multiple linear regression model for heteroscedasticity and linearity. The function outputs orientation plots
and computes relevant statistical tests (heteroscedasticity = lmtest::bptest; linearity = lmtest::raintest).
}
\examples{
data("data_s")

model <- lm(sDur ~ typeOfS + pauseDur, data_s)

heteroscedasticity_check(model)

}
\references{
Achim Zeileis, Torsten Hothorn (2002). Diagnostic Checking in Regression Relationships. R News 2(3), 7-10. URL: https://CRAN.R-project.org/doc/Rnews/
}
\author{
D. Schmitz & J. Esser
}
