% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mode_stat.R
\name{mode_stat}
\alias{mode_stat}
\title{Mode (Statistical)}
\usage{
mode_stat(x)
}
\arguments{
\item{x}{A numeric vector containing the values whose mode is to be computed.}
}
\description{
Compute the mode of a numerical or ordinal variable.
}
\examples{
data("data_s")

mode_stat(x = data_s$age)

}
\author{
D. Schmitz
}
