% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normality_check.R
\name{normality_check}
\alias{normality_check}
\title{Check Normality of a Model}
\usage{
normality_check(mdl)
}
\arguments{
\item{mdl}{A simple, multiple, or mixed-effects linear regression model.}
}
\value{
A plot and a test results in the console.
}
\description{
Checks a simple, multiple, or mixed-effects linear regression model for normality. The function outputs an orientation plot
and computes a relevant statistical test: Shapiro-Wilk test on model residuals.
}
\examples{
data("data_s")

model <- lm(sDur ~ typeOfS + pauseDur, data_s)

normality_check(model)

}
\author{
D. Schmitz & J. Esser
}
