% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_exercise.R
\name{open_exercise}
\alias{open_exercise}
\title{Open Exercise}
\usage{
open_exercise(session, year, edition, language = NULL)
}
\arguments{
\item{year}{The year of the SfL workshop. Currently either \code{2021} or \code{2022}.}

\item{edition}{The edition of the SfL workshop. Either \code{"english"}, \code{"german"}, or \code{"hannover"}.}

\item{language}{Deprecated. Please use \code{edition} instead.}

\item{session_number}{The number of the SfL session.}
}
\value{
Nothing.
}
\description{
Calling \code{open_exercise} with a session number as argument will open the exercise html for that session.
The html will be opened in your default browser. This function is inspired by Coretta et al. (see references).
}
\examples{
open_exercise(02, 2021, "english")

}
\references{
Coretta, S., Casillas, J. V., & Roettger, T. (2021). learnB4SS: Learning materials for the learnB4SS workshop. R package version 1.0.0. https://github.com/learnB4SS/learnB4SS
}
\author{
D. Schmitz
}
