% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_slide.R
\name{open_slide}
\alias{open_slide}
\title{Open Slide}
\usage{
open_slide(session, year, language)
}
\arguments{
\item{year}{The year of the SfL workshop. Currently either \code{2021} or \code{2022}.}

\item{language}{The language of the SfL workshop. Currently either \code{"english"} or \code{"german"}.}

\item{session_number}{The number of the SfL session.}
}
\value{
Nothing.
}
\description{
Calling \code{open_slides} with a session number as argument will open the powerpoint slides for that session as pdf.
The pdf will be opened in your default browser. This function is inspired by Coretta et al. (see references).
}
\examples{
open_slide(02, 2021, "english")

}
\references{
Coretta, S., Casillas, J. V., & Roettger, T. (2021). learnB4SS: Learning materials for the learnB4SS workshop. R package version 1.0.0. https://github.com/learnB4SS/learnB4SS
}
\author{
D. Schmitz
}
