% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordinal_p.R
\name{ordinal_p}
\alias{ordinal_p}
\title{Compute p-values for Ordinal Regression}
\usage{
ordinal_p(mdl)
}
\arguments{
\item{mdl}{The fitted ordinal regression model.}
}
\value{
Usually used without variable assignment.
}
\description{
Computes p-values for an ordinal regression model fitted with MASS::polr.
}
\examples{
data("data_o")

ord_mdl <- polr(size ~ vowel + C1 + C2 + age, data = data_o, Hess=TRUE)

ordinal_p(mdl = ord_mdl)

}
\references{
Venables, W. N. & Ripley, B. D. (2002) Modern Applied Statistics with S. Fourth Edition. Springer, New York.
}
\author{
D. Schmitz
}
