% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictor_competition.R
\name{predictor_competition}
\alias{predictor_competition}
\title{Compare Variables for Predictor Strength}
\usage{
\code{predictor_competition(data = ,
                      dependent = "",
                      independent1 = "",
                      independent2 = "",
                      random.intercept = "",
                      random.slope = 1)}
}
\arguments{
\item{data}{The original data set for both models.}

\item{dependent}{The dependent variable for both models.}

\item{independent1}{The independent variable(s), i.e. the fixed effects, of the 1st model.}

\item{independent2}{The independent variable(s), i.e. the fixed effects, of the 2nd model.}

\item{random.intercept}{The random intercept for both models.}

\item{random.slope}{The random slope for both models. The default assumes no random slope.}
}
\value{
A dataframe containing npar, AIC, BIG, logLik, deviance, Chisq, Df, and Pr(>Chisq). Usually used without variable assignment.
}
\description{
Compare the predictive strength of two independent variables in a minimal linear mixed effects regression model. The function creates two identical
\code{lmer} objects, only differing in fixed effects structure. Then, a log-likelihood test is used to decide which fixed effect structure is better fit to
predict the dependent variable.
}
\examples{
data("data_s")

# example 1: two similarly well fit predictors
predictor_competition(data = data_s, dependent = "sDur", independent1 = "typeOfS", independent2 = "pauseBin", random.intercept = "speaker")

# example 2: one predictor is better than the other
predictor_competition(data = data_s, dependent = "sDur", independent1 = "typeOfS", independent2 = "slideNumber", random.intercept = "speaker")

}
\references{
Bates, Douglas, Maechler, Martin, Bolker, Ben, and Steve Walker. (2015). Fitting Linear Mixed-Effects Models Using lme4. Journal of Statistical Software, 67(1), 1-48. doi:10.18637/jss.v067.i01.
}
\author{
D. Schmitz
}
