% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictor_strength.R
\name{predictor_strength}
\alias{predictor_strength}
\title{Compute Predictor Strength}
\usage{
\code{predictor_strength(dependent = "",
                   fixed = c(""),
                   random_str = c("( | )"),
                   data = )}
}
\arguments{
\item{dependent}{The dependent variable for all models.}

\item{fixed}{The independent variables, i.e. the fixed effects, for all models.}

\item{random_str}{The random effect structure for all models.}

\item{data}{The original data set.}
}
\value{
A dataframe containing the conditional and marginal coefficients of determination for each model, and a column defining the variable the
values belong to.
}
\description{
Given a dependent variable, a set of fixed effect variables, and a random effect structure, this function creates an \code{lmer} model
for each predictor variable, lacking that predictor variable.
Then, conditional and marginal coefficients of determination for each model are calculated. Comparing the value of the conditional coefficient
of determination across all models, one can conclude the predictor strength of the respective missing predictor.
The function uses \code{MuMIn::r.squaredGLMM} to compute coefficients of determination.
}
\examples{
data("data_s")

predictor_strength(dependent = "sDur",
        fixed = c("pauseBin", "list", "folType", "baseDur"),
        random_str = c("(1 | speaker) + (1 | item)"),
        data = data_s)

}
\references{
Bates, Douglas, Maechler, Martin, Bolker, Ben, and Steve Walker. (2015). Fitting Linear Mixed-Effects Models Using lme4. Journal of Statistical Software, 67(1), 1-48. doi:10.18637/jss.v067.i01.

Barton, Kamil. (2020). MuMIn: Multi-Model Inference. R package version 1.43.17. https://CRAN.R-project.org/package=MuMIn

Nakagawa, S., Johnson, P.C.D., and Schielzeth, H. (2017) The coefficient of determination R? and intra-class correlation coefficient from generalized linear mixed-effects models revisited and expanded. J. R. Soc. Interface 14: 20170213.

R Core Team. (2021). R: A language and environment for statistical computing. R Foundation for Statistical Computing, Vienna, Austria. URL https://www.R-project.org/.
}
\author{
D. Schmitz & J. Esser
}
