% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se.R
\name{se}
\alias{se}
\title{Standard Error}
\usage{
se(x)
}
\arguments{
\item{x}{A numeric vector containing the values whose stadard error is to be computed.}
}
\description{
Compute the standard error of a numerical or ordinal variable.
}
\examples{
data("data_s")

se(data = data_s$age)

}
\author{
D. Schmitz
}
