% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tukey.R
\name{tukey}
\alias{tukey}
\title{Tukey Contrasts}
\usage{
\code{tukey(model = , predictor = )}
}
\arguments{
\item{model}{The model object, created by \code{lm()} or \code{lmer()}.}

\item{predictor}{The categorical predictor for which Tukey Contrasts should be computed.}
}
\value{
A list object if assigned to a variable name.
}
\description{
Compute Tukey Contrasts for all levels of a categorical predictor. Takes simple linear regression
models and multiple linear regression models, as well as linear mixed effects models as input.
}
\examples{
data("data_s")

mdl <- lm(sDur ~ pauseBin, data = data_s)

tukey(model = mdl, predictor = pauseBin)

}
\references{
Hothorn, Torsten, Bretz, Frank, and Westfall, Peter. (2008). Simultaneous Inference in General Parametric Models. Biometrical Journal 50(3), 346-363.
}
\author{
D. Schmitz
}
