% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tukey.R
\name{tukey}
\alias{tukey}
\title{Tukey Contrasts}
\usage{
\code{tukey(model = , predictor = )}
}
\arguments{
\item{model}{The model object, created by \code{lm()} or \code{lmer()}.}

\item{predictor}{The categorical predictor for which Tukey Contrasts should be computed.}
}
\value{
A list object if assigned to a variable name.
}
\description{
Compute Tukey Contrasts for all levels of a categorical predictor. Takes simple linear regression
models and multiple linear regression models (\code{lm()}), as well as linear mixed effects models (\code{lmer()}) as input.
}
\author{
D. Schmitz
}
